add async test, move to test_deco
